+++ packages/apollo-server-cache-memcached/src/__tests__/Memcached.test.ts
@@ -2,16 +2,18 @@
 jest.mock('memcached', () => require('memcached-mock'));
 
 import { MemcachedCache } from '../index';
-import {
-  testKeyValueCache_Basics,
-  testKeyValueCache_Expiration,
-} from '../../../apollo-server-caching/src/__tests__/testsuite';
+import { runKeyValueCacheTests } from 'apollo-server-caching';
+import FakeTimers from '@sinonjs/fake-timers';
 
 describe('Memcached', () => {
-  const cache = new MemcachedCache('localhost');
-  afterAll(async () => {
-    await cache.close();
-  })
-  testKeyValueCache_Basics(cache);
-  testKeyValueCache_Expiration(cache);
+  it('run apollo-server-caching test suite', async () => {
+    const cache = new MemcachedCache('localhost');
+    const clock = FakeTimers.install();
+    try {
+      await runKeyValueCacheTests(cache, (ms: number) => clock.tick(ms));
+    } finally {
+      clock.uninstall();
+      await cache.close();
+    }
+  });
 });