+++ packages/apollo-server-caching/src/PrefixingKeyValueCache.ts
@@ -4,12 +4,12 @@ import { KeyValueCache, KeyValueCacheSetOptions } from './KeyValueCache';
 // by all operations.  This allows multiple features to share the same
 // underlying cache without conflicts.
 //
-// Note that PrefixingKeyValueCache explicitly does not implement
-// TestableKeyValueCache, and notably does not implement the flush()
-// method. Most implementations of TestableKeyValueCache.flush() send a simple
-// command that wipes the entire backend cache system, which wouldn't support
-// "only wipe the part of the cache with this prefix", so trying to provide a
-// flush() method here could be confusingly dangerous.
+// Note that PrefixingKeyValueCache explicitly does not implement methods like
+// flush() that aren't part of KeyValueCache, even though most KeyValueCache
+// implementations also have a flush() method. Most implementations of flush()
+// send a simple command that wipes the entire backend cache system, which
+// wouldn't support "only wipe the part of the cache with this prefix", so
+// trying to provide a flush() method here could be confusingly dangerous.
 export class PrefixingKeyValueCache<V = string> implements KeyValueCache<V> {
   constructor(private wrapped: KeyValueCache<V>, private prefix: string) {}