+++ packages/apollo-server-fastify/src/__tests__/fastifyApollo.test.ts
@@ -7,8 +7,6 @@ import testSuite, {
 } from 'apollo-server-integration-testsuite';
 import { Config } from 'apollo-server-core';
 
-const port = 9998;
-
 async function createApp(options: CreateAppOptions = {}) {
   const app = fastify();
 
@@ -17,7 +15,7 @@ async function createApp(options: CreateAppOptions = {}) {
   );
 
   app.register(server.createHandler());
-  await app.listen();
+  await app.listen(0);
 
   return app.server;
 }