+++ packages/apollo-server-micro/src/__tests__/microApollo.test.ts
@@ -3,7 +3,7 @@ import testSuite, {
   schema as Schema,
   CreateAppOptions,
 } from 'apollo-server-integration-testsuite';
-import { GraphQLOptions, Config } from 'apollo-server-core';
+import { Config } from 'apollo-server-core';
 
 import { ApolloServer } from '../ApolloServer';
 
@@ -16,7 +16,7 @@ function createApp(options: CreateAppOptions = {}) {
 
 describe('microApollo', function() {
   it('should throw an error if called without a schema', function() {
-    expect(() => new ApolloServer(undefined as GraphQLOptions)).toThrow(
+    expect(() => new ApolloServer(undefined as any)).toThrow(
       'ApolloServer requires options.',
     );
   });